Refactored screens into separate folders.
