+++ lib/screens/login_register/register.dart
@@ -1,11 +1,10 @@
+import 'package:adoption_hero/screens/login_register/login.dart';
 import 'package:adoption_hero/screens/navigator_scaffold.dart';
+import 'package:adoption_hero/widgets/error_alert.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:adoption_hero/screens/login.dart';
 import 'package:intl/intl.dart';
-
-import '../widgets/error_alert.dart';
